updated code & add postman collection for test

+++ src/main/java/com/workintech/twitter_clone_api/dto/CommentRequest.java
@@ -13,8 +13,8 @@ public class CommentRequest {
     private String content;
 
     @NotNull
-    private Long userId;
+    private Integer userId;
 
     @NotNull
-    private Long tweetId;
+    private Integer tweetId;
 }
