+++ src/main/java/com/workintech/twitter_clone_api/repository/LikeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface LikeRepository extends JpaRepository<Like, Long> {
+public interface LikeRepository extends JpaRepository<Like, Integer> {
     boolean existsByUserAndTweet(User user, Tweet tweet);
     Optional<Like> findByUserAndTweet(User user, Tweet tweet);
 }
