+++ src/main/java/com/workintech/twitter_clone_api/repository/RetweetRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface RetweetRepository extends JpaRepository<Retweet, Long> {
+public interface RetweetRepository extends JpaRepository<Retweet, Integer> {
     boolean existsByUserAndTweet(User user, Tweet tweet);
     Optional<Retweet> findByUserAndTweet(User user, Tweet tweet);
 }
