+++ src/main/java/com/workintech/twitter_clone_api/repository/TweetRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TweetRepository extends JpaRepository<Tweet, Long> {
-    List<Tweet> findByUserId(Long userId);
+public interface TweetRepository extends JpaRepository<Tweet, Integer> {
+    List<Tweet> findByUserId(Integer userId);
 }
